WE-546 Use a single ApiClient
